Everything is named properly now cleaned up code.
