+++ views/allPokemon.pug
