+++ views/individualPokemon.pug
