+++ views/partials/header.pug
@@ -5,6 +5,6 @@ header
     nav
       ul
         li
-          a(href="/users") All Pokemon
+          a(href="/pokemon") All Pokemon
         li
-          a(href="/users/mew") Mew 
+          a(href="/pokemon/mew") Mew
